+++ src/test/java/org/openlmis/stockmanagement/service/JasperReportServiceTest.java
@@ -67,6 +67,7 @@ public class JasperReportServiceTest {
   private static final String TIME_ZONE_ID = "UTC";
   private static final String GROUPING_SEPARATOR = ",";
   private static final String GROUPING_SIZE = "3";
+  private static final String SHOW_IN_DOSES = "showInDoses";
 
   private JasperReportService jasperReportService;
 
@@ -107,7 +108,7 @@ public class JasperReportServiceTest {
     UUID stockCardId = UUID.randomUUID();
     when(stockCardService.findStockCardById(stockCardId)).thenReturn(null);
 
-    jasperReportService.generateStockCardReport(stockCardId, "en");
+    jasperReportService.generateStockCardReport(stockCardId, "en", true);
   }
 
   @Test
@@ -119,9 +120,15 @@ public class JasperReportServiceTest {
     when(reportService.fillAndExportReport(any(String.class), any(byte[].class), anyMap()))
         .thenReturn(testReportData);
 
-    byte[] reportData = jasperReportService.generateStockCardReport(stockCard.getId(), "en");
+    byte[] reportData =
+        jasperReportService.generateStockCardReport(stockCard.getId(), "en", false);
 
     assertEquals(testReportData, reportData);
+
+    verify(reportService).fillAndExportReport(any(String.class), any(byte[].class),
+        paramsCaptor.capture());
+    assertEquals(Boolean.FALSE, paramsCaptor.getValue().get(SHOW_IN_DOSES));
+    assertEquals(1L, paramsCaptor.getValue().get("orderableNetContent"));
   }
 
   @Test
@@ -138,9 +145,13 @@ public class JasperReportServiceTest {
         .thenReturn(testReportData);
 
     byte[] reportData = jasperReportService.generateStockCardSummariesReport(programId,
-        facilityId, "en");
+        facilityId, "en", false);
 
     assertEquals(testReportData, reportData);
+
+    verify(reportService).fillAndExportReport(any(String.class), any(byte[].class),
+        paramsCaptor.capture());
+    assertEquals(Boolean.FALSE, paramsCaptor.getValue().get(SHOW_IN_DOSES));
   }
 
   @Test
@@ -164,7 +175,7 @@ public class JasperReportServiceTest {
     assertEquals(DATE_FORMAT, params.get("dateFormat"));
     assertEquals(DATE_TIME_FORMAT, params.get("dateTimeFormat"));
     assertEquals(createDecimalFormat(), params.get("decimalFormat"));
-    assertEquals(true, params.get("showInDoses"));
+    assertEquals(true, params.get(SHOW_IN_DOSES));
   }
 
   @Test
@@ -178,7 +189,7 @@ public class JasperReportServiceTest {
 
     verify(reportService).fillAndExportReport(eq("stockEvent"), any(byte[].class),
         paramsCaptor.capture());
-    assertEquals(false, paramsCaptor.getValue().get("showInDoses"));
+    assertEquals(false, paramsCaptor.getValue().get(SHOW_IN_DOSES));
   }
 
   @Test
@@ -188,7 +199,7 @@ public class JasperReportServiceTest {
     params.put("dateFormat", DATE_FORMAT);
     params.put("format", "pdf");
     params.put("decimalFormat", createDecimalFormat());
-    params.put("showInDoses", false);
+    params.put(SHOW_IN_DOSES, false);
 
     ByteArrayOutputStream bos = new ByteArrayOutputStream();
     ObjectOutputStream out = new ObjectOutputStream(bos);